s4bio/requirements.txt
numpy
scipy
pandas
scikit-learn
matplotlib
#tqdm
rich
#torchtext
lit # Getting installation errors with torch 2.0 if this isn't installed
# torchvision
pytorch-lightning==2.0.4
hydra-core
omegaconf
wandb
einops
cmake # For pykeops support
# pykeops # Seems to cause various issues; leaving uninstalled by default
transformers # For some schedulers

# Model specific packges
# pytorch-fast-transformers # for Performer

# Dataset specific packages
datasets # LRA
sktime # BIDMC
scikit-learn # Impedance
numba # Impedance
gluonts # Monash
timm==0.5.4 # ImageNet
